Move CompactTarget to primitives
Move the `CompactTarget` type to `primitives`. Requires using
`to_consensus` and `from_consensus` instead of accessing the now private
inner field but both are cheap to call.
